// src/modules/informes/pages/rep-dispensacion/components/ModalEntregaPacientes/types.ts
export interface Props {
  open: boolean;
  setOpen: (value: boolean) => void;
  data: string[];
}

export interface DataType {
  key: React.Key;
  atc: string;
  autorizacion_detalle: string;
  bod_cencosto: string;
  bod_localicad: string;
  bod_nombre: string;
  bod_tercero: string;
  bodega_id: string;
  cadena_frio: string;
  cant_presentacion: string;
  cantidad_dev: string;
  cantidad_entregada: string;
  cantidad_saldo: string;
  cantidad_solicitada: string;
  circular_regulacion: string;
  cod_barra: string;
  cod_geminus: string;
  cod_mipres: string;
  cod_padre: string;
  codigo: string;
  concentracion: string;
  consecutivo: string;
  costo_impuesto: string;
  costo_subtotal: string;
  costo_total: string;
  costo_unitario: string;
  created_at: string;
  cum: string;
  descripcion: string;
  dias_tratamiento: string;
  direccion: string;
  dispensacion_id: string;
  estado: string;
  estado_apertura: string;
  estado_invima: string;
  expediente: string;
  fecha_sync: string;
  fecha_vencimiento: string;
  fecha_vig_invima: string;
  flag_pendientes: string;
  forma_far: string;
  grupo_id: string;
  id: string;
  id_bod_yemi: string;
  id_contrato: string;
  id_empresa: string;
  id_grupo: string;
  id_motivo_detalle: string;
  id_tipo_factu: string;
  invima: string;
  ium: string;
  ium1: string;
  ium2: string;
  ium3: string;
  iva_id: string;
  laboratorio: string;
  lote: string;
  nivel_invima: string;
  nombre: string;
  numero_linea: string;
  p_regulado_compra: string;
  p_regulado_venta: string;
  pbs: string;
  precio_compra: string;
  precio_iva: string;
  precio_promedio: string;
  precio_subtotal: string;
  precio_venta: string;
  precio_venta_total: string;
  prefijo: string;
  presentacion: string;
  principio_act: string;
  producto_id: string;
  producto_lote_id: string;
  stock: string;
  subgrupo_id: string;
  sync: string;
  tabla_bd: string;
  tabla_db: string;
  tipo_bodega: string;
  ubicacion: string;
  uni_dispensacion: string;
  unidad_medida: string;
  updated_at: string;
  usuario_sync: string;
  via_admin: string;
  producto_descripcion: string;
  numero_identificacion: string;
  nombre_primero: string;
  nombre_segundo: string;
  apellido_primero: string;
  apellido_segundo: string;
  observacion: string;
}